% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_collection.R
\name{check_time_series_collection}
\alias{check_time_series_collection}
\title{Check that a collection of time series is properly formatted}
\usage{
check_time_series_collection(collection)
}
\arguments{
\item{collection}{a list representing a collection of time series as
described in \link{plot_collection}.}
}
\value{
a character string with value \code{"OK"} if the object is properly
formatted. Otherwise, the character string indicates the first error found
in the object format.
}
\description{
This function checks that an object holding a collection of time series,
their future values and their forecasts has the correct format. This kind of
objects are used in function \link{plot_collection}.
}
\examples{
c <- list(
   list(Historical = window(USAccDeaths, end = c(1977, 12)),
        Future = window(USAccDeaths, start = c(1978, 1)),
        Forecasts = list(mean = rep(mean(window(USAccDeaths, end = c(1977, 12))), 12),
                         naive = rep(tail(window(USAccDeaths, end = c(1977, 12)), 1), 12)
        )
   ),
   list(Historical = window(UKDriverDeaths, end = c(1983, 12)),
        Future = window(UKDriverDeaths, start = c(1984, 1)),
        Forecasts = list(mean = rep(mean(window(UKDriverDeaths, end = c(1983, 12))), 12),
                         naive = rep(tail(window(UKDriverDeaths, end = c(1983, 12)), 1), 12)
        )
   )
)
check_time_series_collection(c)
}
