% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI_collection.R
\name{GUI_collection}
\alias{GUI_collection}
\title{Launches the web-based GUI for visualizing time series}
\usage{
GUI_collection(collection)
}
\arguments{
\item{collection}{a list with the collection of time series. Each component
of the list must have been built with the \code{\link[=ts_info]{ts_info()}} function.}
}
\value{
Nothing
}
\description{
Launches the web-based GUI for visualizing time series
}
\examples{
# create a collection of two time series and visualize them
c <- list(ts_info(USAccDeaths), ts_info(ldeaths))
\dontrun{
GUI_collection(c)
}
}
