% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_collection.R
\name{plot_collection}
\alias{plot_collection}
\title{Create the associated ggplot object of a time series belonging to a
collection}
\usage{
plot_collection(collection, number, sdp = TRUE)
}
\arguments{
\item{collection}{a list with the collection of time series. Each component
of the list must have been built with the \code{\link[=ts_info]{ts_info()}} function.}

\item{number}{an integer. The number of the time series. It should be a value
between 1 and \code{length(collection)}.}

\item{sdp}{logical. Should data points be shown in the plot? (default value \code{TRUE})}
}
\value{
The ggplot object representing the time series and its forecast.
}
\description{
Create a \code{ggplot} object associated with a time series belonging to a
collection. Apart from the time series, future values and forecasts for the
future values form part of the \code{ggplot} object.
}
\details{
The \code{collection} parameter must be a list. Each component of the list stores
a time series and, optionally, its future values, forecasts for the future
values and prediction intervals for the forecasts. Each component should have
been created using the \code{\link[=ts_info]{ts_info()}} function.

In the example section you can see an example of a collection of time series.
If the \code{collection} parameter is not specified correctly, a proper message is
shown.
}
\examples{
# create a collection of two time series and plot both time series
c <- list(
          ts_info(USAccDeaths),
          ts_info(ldeaths)
)
plot_collection(c, number = 1)
plot_collection(c, number = 2)
}
\seealso{
\code{\link[=ts_info]{ts_info()}} function to see how to build the components of the
\code{collection} parameter.
}
